Add changes from https://github.com/tomchristie/apistar/pull/104
