+++ .travis.yml
@@ -7,7 +7,6 @@ python:
     - "3.6"
 
 env:
-    - DB_URL=sqlite:///test.db
     - DB_URL=postgresql://:@localhost/apistar
 
 install:
