+++ apistar/backends/__init__.py
@@ -1,0 +1,3 @@
+from apistar.backends.sqlalchemy import SQLAlchemy
+
+__all__ = ['SQLAlchemy']
