+++ apistar/commands/__init__.py
@@ -1,4 +1,4 @@
-from apistar.commands.db import create_tables
+from apistar.commands.create_tables import create_tables
 from apistar.commands.new import new
 from apistar.commands.run import run
 from apistar.commands.test import test
