+++ apistar/commands/create_tables.py
@@ -1,14 +1,13 @@
 import click
 
-from apistar.db import DBBackend
-
 
 def create_tables():
     """
     Create SQLAlchemy tables.
     """
     from apistar.main import get_current_app
+    from apistar.backends import SQLAlchemy
     app = get_current_app()
-    db_backend = DBBackend.build(db_config=app.settings.get('DATABASE', {}))
+    db_backend = SQLAlchemy.build(settings=app.settings)
     db_backend.create_tables()
     click.echo("Tables created")
